+++ internal/provider/ephemeral_resource_team_token.go
@@ -5,8 +5,12 @@ package provider
 
 import (
 	"context"
+	"encoding/json"
+	"errors"
 	"fmt"
+	"log"
 
+	"github.com/google/uuid"
 	"github.com/hashicorp/go-tfe"
 	"github.com/hashicorp/terraform-plugin-framework-timetypes/timetypes"
 	"github.com/hashicorp/terraform-plugin-framework/ephemeral"
@@ -28,15 +32,25 @@ type TeamTokenEphemeralResource struct {
 }
 
 type TeamTokenEphemeralResourceModel struct {
+	ID        types.String      `tfsdk:"id"`
 	TeamID    types.String      `tfsdk:"team_id"`
 	Token     types.String      `tfsdk:"token"`
 	ExpiredAt timetypes.RFC3339 `tfsdk:"expired_at"`
+}
+
+type TeamTokenPrivateData struct {
+	ID     string `json:"id"`
+	TeamID string `json:"team_id"`
 }
 
 func (e *TeamTokenEphemeralResource) Schema(ctx context.Context, req ephemeral.SchemaRequest, resp *ephemeral.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		Description: "This ephemeral resource can be used to retrieve a team token without saving its value in state.",
 		Attributes: map[string]schema.Attribute{
+			"id": schema.StringAttribute{
+				Description: `System generated ID of the team token.`,
+				Computed:    true,
+			},
 			"team_id": schema.StringAttribute{
 				Description: `ID of the team.`,
 				Required:    true,
@@ -88,8 +102,13 @@ func (e *TeamTokenEphemeralResource) Open(ctx context.Context, req ephemeral.Ope
 		return
 	}
 
+	desc := fmt.Sprintf("ephemeral-team-token-%s", uuid.New())
+
 	// Create a new options struct
-	options := tfe.TeamTokenCreateOptions{}
+	// Set a description to make use of the new multiple token API
+	options := tfe.TeamTokenCreateOptions{
+		Description: &desc,
+	}
 
 	if !config.ExpiredAt.IsNull() {
 		expiredAt, diags := config.ExpiredAt.ValueRFC3339Time()
@@ -108,10 +127,37 @@ func (e *TeamTokenEphemeralResource) Open(ctx context.Context, req ephemeral.Ope
 		return
 	}
 
+	privateData, _ := json.Marshal(result)
+	resp.Private.SetKey(ctx, "team_token_data", privateData)
+
 	config = ephemeralResourceModelFromTFETeamToken(teamID, result)
 
 	// Save to ephemeral result data
 	resp.Diagnostics.Append(resp.Result.Set(ctx, &config)...)
+}
+
+func (e *TeamTokenEphemeralResource) Close(ctx context.Context, req ephemeral.CloseRequest, resp *ephemeral.CloseResponse) {
+	privateBytes, diags := req.Private.GetKey(ctx, "team_token_data")
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	var privateData TeamTokenPrivateData
+	json.Unmarshal(privateBytes, &privateData)
+
+	log.Printf("[DEBUG] Removing team token with ID: %s", privateData.ID)
+
+	err := e.config.Client.TeamTokens.Delete(ctx, privateData.ID)
+	if err != nil {
+		if errors.Is(err, tfe.ErrResourceNotFound) {
+			log.Printf("[DEBUG] Team token with ID %s not found, skipping deletion", privateData.ID)
+			return
+		}
+
+		resp.Diagnostics.AddError("Unable to delete team token", err.Error())
+		return
+	}
 }
 
 // ephemeralResourceModelFromTFETeamToken builds a TeamTokenEphemeralResourceModel struct from a